% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myData.R
\docType{data}
\name{chlam_df}
\alias{chlam_df}
\title{chlam_df: chlamydia infection data}
\format{
A data frame with 1251 rows and 12 variables:
\describe{
  \item{age}{Age at screening, scaled}
  \item{nonhisp.white}{Non Hispanic White}
  \item{nonhisp.black}{Non Hispanic Black}
  \item{mexhisp}{Mexican American and Hispanic}
  \item{pvrtyind}{Family monthly poverty level index, scaled}
  \item{newsexp}{New sex partner in the past 12 months}
  \item{pvrtying.geq5}{Poverty index > 5}
  \item{gender}{Gender male 1 female 0}
  \item{edughigh}{Education greater than high school}
  \item{nmarried}{never married}
}
}
\usage{
data(chlam_df)
}
\description{
This dataset contains chlamydia infection data collected through the National Health and Nutrition Examination Survey from NHANES
}
\keyword{datasets}
