% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{E.Y}
\alias{E.Y}
\title{Function to calculate conditional expectation of Y given observed response Z for pool Pj}
\usage{
E.Y(Se, Sp, pi, Z)
}
\arguments{
\item{Se}{sensitivity}

\item{Sp}{specificity}

\item{pi}{a vector of probabilities of length mj, number of samples within jth pool}

\item{Z}{observed group response either 0 or 1}
}
\value{
numeric vector of expected value of Y given group response Z
}
\description{
Function to calculate conditional expectation of Y given observed response Z for pool Pj
}
\examples{
# generate some data
n <- 1000
p <- 5
n_groups <- 250
m <- n/n_groups
Se <- 0.95
Sp <- 0.95
X <- cbind(rep(1,n),matrix(rnorm(n*p),nrow=n))
beta_hat <- c(-5,3,-2,1,0,0) # first is intercept
pi.hat <- as.numeric(logistic(X \%*\% beta_hat))

E.Y(Se=0.95, Sp=0.95, pi=pi.hat[1:m], Z=0) #first pool
}
