% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM_group}
\alias{EM_group}
\title{Function to perform EM algorithm for group testing}
\usage{
EM_group(X, Z, Se, Sp, lambda, tol = 0.001, n_groups)
}
\arguments{
\item{X}{design matrix. First column is treated as an intercept.}

\item{Z}{observed group response vector of 0s and 1s}

\item{Se}{sensitivity}

\item{Sp}{specificity}

\item{lambda}{the lasso sparsity tuning parameter}

\item{tol}{convergence tolerance}
}
\value{
numeric vector of estimated regression coefficients beta_hat, and estimated response Y.hat
}
\description{
Function to perform EM algorithm for group testing
}
\examples{
# generate some data
n <- 1000
p <- 5
Se <- 0.95
Sp <- 0.95
beta_true <- c(-5,3,0,1,-2,0) # first is intercept
X <- cbind(rep(1,n),matrix(rnorm(n*p),nrow=n))
pi.true <- as.numeric(logistic(X \%*\% beta_true))
Y <- rbinom(n,1,prob = pi.true)
n_groups <- 250
m <- n/n_groups
Z <- c()
for (j in 1:n_groups){
  maxY <- max(Y[((j-1)*m+1):(j*m)])
  Z[j] <- rbinom(1,1,prob = Se*maxY+(1-Sp)*(1-maxY))
}

EM_group(X=X, Z=Z, Se=0.95, Sp=0.95, lambda=5, tol=1e-3)
}
