% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM_individial}
\alias{EM_individial}
\title{Function to perform EM algorithm for individual testing}
\usage{
EM_individial(X, Z, Se, Sp, lambda, tol = 0.001)
}
\arguments{
\item{X}{design matrix. First column is treated as an intercept.}

\item{Z}{observed individual response vector of 0s and 1s}

\item{Se}{sensitivity}

\item{Sp}{specificity}

\item{lambda}{the lasso sparsity tuning parameter}

\item{tol}{convergence tolerance}
}
\value{
numeric vector of estimated regression coefficients beta_hat, and estimated response Y.hat
}
\description{
Function to perform EM algorithm for individual testing
}
\examples{
# generate some data
n <- 1000
p <- 5
Se <- 0.95
Sp <- 0.95
beta_true <- c(0,3,-2,1,0,0) # first is intercept
X <- cbind(rep(1,n),matrix(rnorm(n*p),nrow=n))
pi.true <- as.numeric(logistic(X \%*\% beta_true))
Y <- rbinom(n,1,prob = pi.true)
Z <- rbinom(n,1,prob = Se*Y+(1-Sp)*(1-Y))

EM_individial(X=X, Z=Z, Se=0.95, Sp=0.95, lambda=5, tol=1e-3)
}
