% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGeneration.R
\name{generate.data.group}
\alias{generate.data.group}
\title{Function to generate group testing data}
\usage{
generate.data.group(n, p, Se, Sp, beta_true, n_groups)
}
\arguments{
\item{n}{number of observations of X}

\item{p}{number of variables of X}

\item{Se}{sensitivity}

\item{Sp}{specificity}

\item{beta_true}{true regression coefficient}

\item{n_groups}{number of groups}
}
\value{
A list of covariates X and observed responses Z
}
\description{
Function to generate group testing data
}
