% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGeneration.R
\name{generate.data}
\alias{generate.data}
\title{Function to generate individual testing data}
\usage{
generate.data(n, p, Se, Sp, beta_true)
}
\arguments{
\item{n}{number of observations of X}

\item{p}{number of variables of X}

\item{Se}{sensitivity}

\item{Sp}{specificity}

\item{beta_true}{true regression coefficient}
}
\value{
A list of covariates X and observed responses Z
}
\description{
Function to generate individual testing data
}
