% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logregLASSO.R
\name{logistic}
\alias{logistic}
\title{Function to compute logistic transformation}
\usage{
logistic(x)
}
\arguments{
\item{x}{any real number}
}
\value{
a number between 0 and 1
}
\description{
Function to compute logistic transformation
}
