% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{loglike_group}
\alias{loglike_group}
\title{Log-likelihood function for logistic regression after lasso selection, group testing}
\usage{
loglike_group(Z, Se, Sp, pi_hat, n_groups, m)
}
\arguments{
\item{Z}{observed group response vector of 0s and 1s}

\item{Se}{sensitivity}

\item{Sp}{specificity}

\item{pi_hat}{a vector of estimated probabilities of length mj}

\item{n_groups}{number of groups}
}
\value{
log likelihood
}
\description{
Log-likelihood function for logistic regression after lasso selection, group testing
}
