% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{loglike_ind}
\alias{loglike_ind}
\title{Log-likelihood function for logistic regression after lasso selection, individual testing}
\usage{
loglike_ind(Z, Se, Sp, pi_hat)
}
\arguments{
\item{Z}{observed group response vector of 0s and 1s}

\item{Se}{sensitivity}

\item{Sp}{specificity}

\item{pi_hat}{a vector of estimated probabilities of length mj}
}
\value{
log likelihood
}
\description{
Log-likelihood function for logistic regression after lasso selection, individual testing
}
