% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logregLASSO.R
\name{logreg_lasso}
\alias{logreg_lasso}
\title{Function to compute logistic regression parameters with a lasso penalty using IRLS}
\usage{
logreg_lasso(X, Y, lambda, tol = 1e-05)
}
\arguments{
\item{X}{design matrix. First column is treated as an intercept.}

\item{Y}{response vector of 0s and 1s}

\item{lambda}{the lasso sparsity tuning parameter}

\item{tol}{convergence tolerance}
}
\value{
numeric vector of estimated regression coefficients
}
\description{
Function to compute logistic regression parameters with a lasso penalty using IRLS
}
\examples{
# generate some data
n <- 1000
p <- 5
b_true <- c(0,3,-2,1,0,0) # first is intercept
X <- cbind(rep(1,n),matrix(rnorm(n*p),nrow=n))
eta <- X \%*\% b_true
Y <- rbinom(n,1,logistic(eta))

logreg_lasso(X,Y,lambda = 5)
}
