% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postInference.R
\name{naive_logreg_group}
\alias{naive_logreg_group}
\title{Classical inference for logistic regression after lasso selection, group testing}
\usage{
naive_logreg_group(X.M, Z, Se, Sp, beta_hat, level.alpha = 0.05, n_groups)
}
\arguments{
\item{X.M}{design matrix with selected columns. First column is treated as an intercept.}

\item{Z}{observed group response vector of 0s and 1s}

\item{Se}{sensitivity}

\item{Sp}{specificity}

\item{beta_hat}{EM estimated coefficient vector}

\item{level.alpha}{the level of confidence}

\item{n_groups}{number of groups}
}
\value{
A list with the Naive inference confidence intervals and p values
}
\description{
Classical inference for logistic regression after lasso selection, group testing
}
