% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postInference.R
\name{naive_logreg_individual}
\alias{naive_logreg_individual}
\title{Classical inference for logistic regression after lasso selection, individual testing}
\usage{
naive_logreg_individual(X.M, Y.hat, beta_hat, level.alpha = 0.05)
}
\arguments{
\item{X.M}{design matrix with selected columns. First column is treated as an intercept.}

\item{Y.hat}{EM estimated response vector of 0s and 1s}

\item{beta_hat}{EM estimated coefficient vector}

\item{level.alpha}{the level of confidence}
}
\value{
A list with the Naive inference confidence intervals and p values
}
\description{
Classical inference for logistic regression after lasso selection, individual testing
}
