% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postInference.R
\name{post_logreg_group}
\alias{post_logreg_group}
\title{Group testing inference for logistic regression after lasso selection}
\usage{
post_logreg_group(
  X,
  Z,
  Y.hat,
  Se,
  Sp,
  lambda,
  beta_hat,
  n_groups,
  level.alpha = 0.05
)
}
\arguments{
\item{X}{design matrix. First column is treated as an intercept.}

\item{Z}{observed group response vector of 0s and 1s}

\item{Y.hat}{EM estimated response vector of 0s and 1s}

\item{Se}{sensitivity}

\item{Sp}{specificity}

\item{lambda}{the lasso sparsity tuning parameter}

\item{beta_hat}{EM estimated coefficient vector}

\item{n_groups}{number of groups}

\item{level.alpha}{the level of confidence}
}
\value{
A list with the post selection confidence intervals, coefficients estimates, Truncated Normal standard deviations, and Truncated Normal limits
}
\description{
Group testing inference for logistic regression after lasso selection
}
\examples{
# generate some data
n <- 1000
p <- 5
Se <- 0.95
Sp <- 0.95
b_true <- c(-5,3,0,1,-2,0) # first is intercept
X <- cbind(rep(1,n),scale(matrix(rnorm(n*p),nrow=n),TRUE,TRUE))
pi.true <- as.numeric(logistic(X \%*\% b_true))
Y <- rbinom(n,1,prob = pi.true)

n_groups <- 250
m <- n/n_groups #number of samples within a group
Z <- c()
for(j in 1:n_groups){
  maxY <- max(Y[((j-1)*m+1):(j*m)])
  Z[j] <- rbinom(1,1,prob=Se*maxY+(1-Sp)*(1-maxY))
}

EM_result <- EM_group(X=X, Z=Z, Se=0.95, Sp=0.95, lambda=5, tol=1e-3, n_groups=250)
beta_hat <- EM_result$beta_hat
Y.hat <- EM_result$Y.hat
post_logreg_group(X,Z,Y.hat,Se,Sp,lambda=5,beta_hat,n_groups,level.alpha = 0.05)
}
