% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postInference.R
\name{post_logreg_individual}
\alias{post_logreg_individual}
\title{Individual testing inference for logistic regression after lasso selection}
\usage{
post_logreg_individual(X, Y.hat, lambda, beta_hat, level.alpha = 0.05)
}
\arguments{
\item{X}{design matrix. First column is treated as an intercept.}

\item{Y.hat}{EM estimated response vector Y of 0s and 1s}

\item{lambda}{the lasso sparsity tuning parameter}

\item{beta_hat}{EM estimated coefficient vector}

\item{level.alpha}{the level of confidence}
}
\value{
A list with the post selection confidence intervals, coefficients estimates, Truncated Normal standard deviations, and Truncated Normal limits
}
\description{
Individual testing inference for logistic regression after lasso selection
}
\examples{
# generate some data
n <- 1000
p <- 5
Se <- 0.95
Sp <- 0.95
b_true <- c(-5,3,0,1,-2,0) # first is intercept
X <- cbind(rep(1,n),scale(matrix(rnorm(n*p),nrow=n),T,T))

pi.true <- as.numeric(logistic(X \%*\% beta_true))
Y <- rbinom(n,1,logistic(pi.true))
Z <- rbinom(n,1,prob = Se*Y+(1-Sp)*(1-Y))
beta_hat <- EM_individial(X=X, Z=Z, Se=0.95, Sp=0.95, lambda=5, tol=1e-3)$beta_hat
Y.hat <- EM_individial(X=X, Z=Z, Se=0.95, Sp=0.95, lambda=5, tol=1e-3)$Y.hat
post_logreg_individual(X,Y.hat,lambda,beta_hat,level.alpha = 0.10)
}
