% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGeneration.R
\name{generate.data}
\alias{generate.data}
\title{Function to generate individual testing data}
\usage{
generate.data(n, p, Se, Sp, beta_true)
}
\arguments{
\item{n}{number of observations of X}

\item{p}{number of variables of X}

\item{Se}{sensitivity}

\item{Sp}{specificity}

\item{beta_true}{true regression coefficient}
}
\value{
A list of covariates X and observed responses Z
}
\description{
Function to generate individual testing data
}
\examples{
# generate individual testing data
n <- 1000
p <- 5
Se <- 0.95
Sp <- 0.95
beta_true <- c(0,3,-2,1,0,0) # first is intercept
generate.data(n, p, Se, Sp, beta_true)
}
